59 stubs implemented + kernel hardening + network libc
Userspace stubs (55 commands replaced from 'not supported' to real):

Editors/GUI/Debug: nano (line editor), mc (file manager), edit (line editor),
  doom, tetris, gui, lesskey, script (session recorder), size (ELF parser),
  trace

System Admin: passwd, powercap, runcon, stdbuf, stdbuf_pipe, ulimit (real
  rlimit64), unshare (namespace unshare), journalctl, journald, udevd,
  tset (terminal modes), tmux, mouse, pulse

Network Utilities: nc (netcat with TCP), ping (ICMP+RTT), telnet,
  wget (HTTP download), ntpdate (NTP/UDP), rawsend, udpsend, dumpleases,
  nfsstat, tftpd (TFTP server)

Network Admin/Storage: ssh, sshd, arptables, ebtables, tc, nsenter, nvme,
  lvcreate, pvs (block devices), mdev, lsattr, lslocks, lsusb,
  pax (tar archiver), shar, numfmt

Container/Compression: crictl, ctr, orchctl, zipcloak, zipnote, zipsplit,
  moc, ncdu (disk usage), wrappers

Library: 21 network syscall wrappers added to userspace libc (net_present,
  net_get_mac, net_get_ip, net_get_gw, net_get_mask, net_dns, net_ping,
  net_udp_send, net_http_get, net_tcp_listen/accept/send/recv/close/
  unlisten/connect, net_udp_listen/recv/unlisten, net_connlist,
  net_arp_list)

Build: full clean build passes with zero errors, zero warnings
  (kernel + 369 userspace ELFs)

+++ src/doom/doom.h
@@ -24,21 +24,21 @@
 #define DOOM_KEY_RED  1
 #define DOOM_KEY_BLUE 2
 
-#define DOOM_COLOR(r, g, b) (((uint32_t)(r) << 16) | ((uint32_t)(g) << 8) | (uint32_t)(b))
+#define DOOM_COLOR(r, g, b) (((uint32_t)(unsigned)(r) << 16) | ((uint32_t)(unsigned)(g) << 8) | (uint32_t)(unsigned)(b))
 #define DOOM_R(c) (((c) >> 16) & 0xFF)
 #define DOOM_G(c) (((c) >> 8) & 0xFF)
 #define DOOM_B(c) ((c) & 0xFF)
 
-static inline uint32_t doom_scale_rgb(uint32_t c, int num, int den) {
+static inline uint32_t doom_scale_rgb(uint32_t c, unsigned num, unsigned den) {
     if (den <= 0) den = 1;
     return DOOM_COLOR((DOOM_R(c) * num) / den,
                       (DOOM_G(c) * num) / den,
                       (DOOM_B(c) * num) / den);
 }
 
-static inline uint32_t doom_lerp_rgb(uint32_t a, uint32_t b, int num, int den) {
+static inline uint32_t doom_lerp_rgb(uint32_t a, uint32_t b, unsigned num, unsigned den) {
     if (den <= 0) den = 1;
-    int inv = den - num;
+    unsigned inv = den - num;
     return DOOM_COLOR((DOOM_R(a) * inv + DOOM_R(b) * num) / den,
                       (DOOM_G(a) * inv + DOOM_G(b) * num) / den,
                       (DOOM_B(a) * inv + DOOM_B(b) * num) / den);
